refactor(ARCO-304): move api mocks to internal
