+++ internal/broadcaster/api_test.go
@@ -11,9 +11,9 @@ import (
 
 	sdkTx "github.com/bitcoin-sv/go-sdk/transaction"
 
+	"github.com/bitcoin-sv/arc/internal/api/mocks"
 	"github.com/bitcoin-sv/arc/internal/metamorph/metamorph_api"
 	"github.com/bitcoin-sv/arc/pkg/api"
-	"github.com/bitcoin-sv/arc/pkg/api/mocks"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
